% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{aseR}
\alias{aseR}
\title{aseR}
\usage{
aseR(nA, nTotal, zeta, maxIt = 50, trace = 0)
}
\arguments{
\item{nA}{nA}

\item{nTotal}{nTotal}

\item{zeta}{zeta}

\item{maxIt}{maxIt}

\item{trace}{trace}
}
\description{
aseR
}
