% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{grad.bxj}
\alias{grad.bxj}
\title{grad.bxj}
\usage{
grad.bxj(bxj, y, x, nA, nTotal, zeta, mu, b0, phi, theta)
}
\arguments{
\item{bxj}{bxj}

\item{y}{y}

\item{x}{x}

\item{nA, }{nA}

\item{nTotal}{nTotal}

\item{zeta}{zeta}

\item{mu}{mu}

\item{b0}{b0}

\item{phi}{phi}

\item{theta}{theta}
}
\description{
grad.bxj
}
