% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{gradLogH1}
\alias{gradLogH1}
\title{gradLogH1}
\usage{
gradLogH1(par, nA, nTotal, zeta)
}
\arguments{
\item{par}{par}

\item{nA}{nA}

\item{nTotal}{nTotal}

\item{zeta}{zeta}
}
\description{
gradLogH1
}
