% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{loglikTheta}
\alias{loglikTheta}
\title{loglikTheta}
\usage{
loglikTheta(theta, pi, nA, nTotal, zeta)
}
\arguments{
\item{theta}{theta}

\item{pi}{pi}

\item{nA, }{nA}

\item{nTotal}{nTotal}

\item{zeta}{zeta}
}
\description{
loglikTheta
}
