% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{logLTReC}
\alias{logLTReC}
\title{logLTReC}
\usage{
logLTReC(bxj, y, x, mu, b0, phi, fam)
}
\arguments{
\item{bxj}{bxj}

\item{y}{y}

\item{x}{x}

\item{mu}{mu}

\item{b0}{b0}

\item{phi}{phi}

\item{fam}{fam}
}
\description{
logLTReC
}
