% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{trecaseR}
\alias{trecaseR}
\title{trecaseR}
\usage{
trecaseR(y, y1, y2, X, z1, z2, plotIt = FALSE, traceIt = FALSE)
}
\arguments{
\item{y}{y}

\item{y1}{y1}

\item{y2}{y2}

\item{z1}{z1}

\item{z2}{z2}

\item{plotIt}{plotIt}

\item{traceIt}{traceIt}

\item{x}{x}
}
\description{
trecaseR
}
