% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_funcs.R
\name{trecR}
\alias{trecR}
\title{trecR}
\usage{
trecR(y, X, z1, fam, nIter = 100, plotIt = FALSE, trace = FALSE,
  yfit = FALSE)
}
\arguments{
\item{y}{y}

\item{z1}{z1}

\item{fam}{fam}

\item{nIter}{nIter}

\item{plotIt}{plotIt}

\item{trace}{trace}

\item{yfit}{yfit}

\item{x}{x}
}
\description{
trecR
}
