% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerASEeQTL.R
\name{powerASEeQTL}
\alias{powerASEeQTL}
\title{powerASEeQTL}
\usage{
powerASEeQTL(n, mu = 500, n.simu = 200, sim = "sim1",
  methods = c("eQTL.linear", "eQTL.negBin", "eQTL.quasipoisson", "ASE",
  "eQTL.ASE"), folds = seq(1, 2, length.out = 3), alpha = 0.001,
  phi = 1.1, theta = 0.1, maf = 0.2, propASE = 0.005,
  legend = TRUE, color = TRUE, title = "", subtitle = "",
  titlecolor = "black", subtitlecolor = "black", titlesize = 1,
  subtitlesize = 1, labelsize = 1, labelcolor = "black",
  linewidth = 2, tilt = 0, SEED = 1)
}
\arguments{
\item{n}{n}

\item{mu}{mu}

\item{n.simu}{n.simu}

\item{sim}{sim}

\item{methods}{methods}

\item{folds}{folds}

\item{alpha}{alpha}

\item{phi}{phi}

\item{theta}{theta}

\item{maf}{maf}

\item{propASE}{propASE}

\item{legend}{legend}

\item{color}{color}

\item{title}{title}

\item{subtitle}{subtitle}

\item{titlecolor}{titlecolor}

\item{subtitlecolor}{subtitlecolor}

\item{titlesize}{titlesize}

\item{subtitlesize}{subtitlesize}

\item{labelsize}{labelsize}

\item{labelcolor}{labelcolor}

\item{linewidth}{linewidth}
}
\value{
matResultsF
}
\description{
A function to
}
\author{
Annie Thwing, Sharon Lutz, Michael Gooch
}
